requirements.txt
Flask>=2.0.0
Flask-CORS>=3.0.0
opencv-python>=4.5.0
requests>=2.25.0
paho-mqtt>=1.5.0
python-dotenv>=0.19.0
Pillow>=8.0.0
numpy>=1.20.0
